+++ sdk/resourcemanager/cosmos/armcosmos/models.go
@@ -3242,8 +3242,14 @@ type MongoDBCollectionGetPropertiesResource struct {
 	// Analytical TTL.
 	AnalyticalStorageTTL *int32 `json:"analyticalStorageTtl,omitempty"`
 
+	// Enum to indicate the mode of resource creation.
+	CreateMode *CreateMode `json:"createMode,omitempty"`
+
 	// List of index keys
 	Indexes []*MongoIndex `json:"indexes,omitempty"`
+
+	// Parameters to indicate the information about the restore
+	RestoreParameters *ResourceRestoreParameters `json:"restoreParameters,omitempty"`
 
 	// A key-value pair of shard keys to be applied for the request.
 	ShardKey map[string]*string `json:"shardKey,omitempty"`
@@ -3300,8 +3306,14 @@ type MongoDBCollectionResource struct {
 	// Analytical TTL.
 	AnalyticalStorageTTL *int32 `json:"analyticalStorageTtl,omitempty"`
 
+	// Enum to indicate the mode of resource creation.
+	CreateMode *CreateMode `json:"createMode,omitempty"`
+
 	// List of index keys
 	Indexes []*MongoIndex `json:"indexes,omitempty"`
+
+	// Parameters to indicate the information about the restore
+	RestoreParameters *ResourceRestoreParameters `json:"restoreParameters,omitempty"`
 
 	// A key-value pair of shard keys to be applied for the request.
 	ShardKey map[string]*string `json:"shardKey,omitempty"`
@@ -3363,6 +3375,12 @@ type MongoDBDatabaseGetPropertiesResource struct {
 	// REQUIRED; Name of the Cosmos DB MongoDB database
 	ID *string `json:"id,omitempty"`
 
+	// Enum to indicate the mode of resource creation.
+	CreateMode *CreateMode `json:"createMode,omitempty"`
+
+	// Parameters to indicate the information about the restore
+	RestoreParameters *ResourceRestoreParameters `json:"restoreParameters,omitempty"`
+
 	// READ-ONLY; A system generated property representing the resource etag required for optimistic concurrency control.
 	Etag *string `json:"_etag,omitempty" azure:"ro"`
 
@@ -3411,6 +3429,12 @@ type MongoDBDatabaseListResult struct {
 type MongoDBDatabaseResource struct {
 	// REQUIRED; Name of the Cosmos DB MongoDB database
 	ID *string `json:"id,omitempty"`
+
+	// Enum to indicate the mode of resource creation.
+	CreateMode *CreateMode `json:"createMode,omitempty"`
+
+	// Parameters to indicate the information about the restore
+	RestoreParameters *ResourceRestoreParameters `json:"restoreParameters,omitempty"`
 }
 
 // MongoDBResourcesClientBeginCreateUpdateMongoDBCollectionOptions contains the optional parameters for the MongoDBResourcesClient.BeginCreateUpdateMongoDBCollection
@@ -3514,6 +3538,20 @@ type MongoDBResourcesClientBeginMongoDBContainerRedistributeThroughputOptions st
 // MongoDBResourcesClientBeginMongoDBContainerRetrieveThroughputDistributionOptions contains the optional parameters for the
 // MongoDBResourcesClient.BeginMongoDBContainerRetrieveThroughputDistribution method.
 type MongoDBResourcesClientBeginMongoDBContainerRetrieveThroughputDistributionOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
+// MongoDBResourcesClientBeginMongoDBDatabaseRedistributeThroughputOptions contains the optional parameters for the MongoDBResourcesClient.BeginMongoDBDatabaseRedistributeThroughput
+// method.
+type MongoDBResourcesClientBeginMongoDBDatabaseRedistributeThroughputOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
+// MongoDBResourcesClientBeginMongoDBDatabaseRetrieveThroughputDistributionOptions contains the optional parameters for the
+// MongoDBResourcesClient.BeginMongoDBDatabaseRetrieveThroughputDistribution method.
+type MongoDBResourcesClientBeginMongoDBDatabaseRetrieveThroughputDistributionOptions struct {
 	// Resumes the LRO from the provided token.
 	ResumeToken string
 }
@@ -4373,6 +4411,16 @@ type Resource struct {
 	Type *string `json:"type,omitempty" azure:"ro"`
 }
 
+// ResourceRestoreParameters - Parameters to indicate the information about the restore.
+type ResourceRestoreParameters struct {
+	// The id of the restorable database account from which the restore has to be initiated. For example:
+	// /subscriptions/{subscriptionId}/providers/Microsoft.DocumentDB/locations/{location}/restorableDatabaseAccounts/{restorableDatabaseAccountName}
+	RestoreSource *string `json:"restoreSource,omitempty"`
+
+	// Time to which the account has to be restored (ISO-8601 format).
+	RestoreTimestampInUTC *time.Time `json:"restoreTimestampInUtc,omitempty"`
+}
+
 // RestorableDatabaseAccountGetResult - A Azure Cosmos DB restorable database account.
 type RestorableDatabaseAccountGetResult struct {
 	// The location of the resource group to which the resource belongs.
@@ -4790,6 +4838,9 @@ type RestorableSQLContainerPropertiesResourceContainer struct {
 	// The conflict resolution policy for the container.
 	ConflictResolutionPolicy *ConflictResolutionPolicy `json:"conflictResolutionPolicy,omitempty"`
 
+	// Enum to indicate the mode of resource creation.
+	CreateMode *CreateMode `json:"createMode,omitempty"`
+
 	// Default time to live
 	DefaultTTL *int32 `json:"defaultTtl,omitempty"`
 
@@ -4798,6 +4849,9 @@ type RestorableSQLContainerPropertiesResourceContainer struct {
 
 	// The configuration of the partition key to be used for partitioning data into multiple partitions
 	PartitionKey *ContainerPartitionKey `json:"partitionKey,omitempty"`
+
+	// Parameters to indicate the information about the restore
+	RestoreParameters *ResourceRestoreParameters `json:"restoreParameters,omitempty"`
 
 	// The unique key policy configuration for specifying uniqueness constraints on documents in the collection in the Azure Cosmos
 	// DB service.
@@ -4878,6 +4932,12 @@ type RestorableSQLDatabasePropertiesResource struct {
 type RestorableSQLDatabasePropertiesResourceDatabase struct {
 	// REQUIRED; Name of the Cosmos DB SQL database
 	ID *string `json:"id,omitempty"`
+
+	// Enum to indicate the mode of resource creation.
+	CreateMode *CreateMode `json:"createMode,omitempty"`
+
+	// Parameters to indicate the information about the restore
+	RestoreParameters *ResourceRestoreParameters `json:"restoreParameters,omitempty"`
 
 	// READ-ONLY; A system generated property that specified the addressable path of the collections resource.
 	Colls *string `json:"_colls,omitempty" azure:"ro"`
@@ -5043,6 +5103,16 @@ type RestoreParameters struct {
 	TablesToRestore []*string `json:"tablesToRestore,omitempty"`
 }
 
+// RestoreParametersBase - Parameters to indicate the information about the restore.
+type RestoreParametersBase struct {
+	// The id of the restorable database account from which the restore has to be initiated. For example:
+	// /subscriptions/{subscriptionId}/providers/Microsoft.DocumentDB/locations/{location}/restorableDatabaseAccounts/{restorableDatabaseAccountName}
+	RestoreSource *string `json:"restoreSource,omitempty"`
+
+	// Time to which the account has to be restored (ISO-8601 format).
+	RestoreTimestampInUTC *time.Time `json:"restoreTimestampInUtc,omitempty"`
+}
+
 // RetrieveThroughputParameters - Cosmos DB retrieve throughput parameters object
 type RetrieveThroughputParameters struct {
 	// REQUIRED; Properties to retrieve throughput parameters object
@@ -5157,6 +5227,9 @@ type SQLContainerGetPropertiesResource struct {
 	// The conflict resolution policy for the container.
 	ConflictResolutionPolicy *ConflictResolutionPolicy `json:"conflictResolutionPolicy,omitempty"`
 
+	// Enum to indicate the mode of resource creation.
+	CreateMode *CreateMode `json:"createMode,omitempty"`
+
 	// Default time to live
 	DefaultTTL *int32 `json:"defaultTtl,omitempty"`
 
@@ -5165,6 +5238,9 @@ type SQLContainerGetPropertiesResource struct {
 
 	// The configuration of the partition key to be used for partitioning data into multiple partitions
 	PartitionKey *ContainerPartitionKey `json:"partitionKey,omitempty"`
+
+	// Parameters to indicate the information about the restore
+	RestoreParameters *ResourceRestoreParameters `json:"restoreParameters,omitempty"`
 
 	// The unique key policy configuration for specifying uniqueness constraints on documents in the collection in the Azure Cosmos
 	// DB service.
@@ -5228,6 +5304,9 @@ type SQLContainerResource struct {
 	// The conflict resolution policy for the container.
 	ConflictResolutionPolicy *ConflictResolutionPolicy `json:"conflictResolutionPolicy,omitempty"`
 
+	// Enum to indicate the mode of resource creation.
+	CreateMode *CreateMode `json:"createMode,omitempty"`
+
 	// Default time to live
 	DefaultTTL *int32 `json:"defaultTtl,omitempty"`
 
@@ -5236,6 +5315,9 @@ type SQLContainerResource struct {
 
 	// The configuration of the partition key to be used for partitioning data into multiple partitions
 	PartitionKey *ContainerPartitionKey `json:"partitionKey,omitempty"`
+
+	// Parameters to indicate the information about the restore
+	RestoreParameters *ResourceRestoreParameters `json:"restoreParameters,omitempty"`
 
 	// The unique key policy configuration for specifying uniqueness constraints on documents in the collection in the Azure Cosmos
 	// DB service.
@@ -5320,6 +5402,12 @@ type SQLDatabaseGetPropertiesResource struct {
 	// A system generated property that specified the addressable path of the collections resource.
 	Colls *string `json:"_colls,omitempty"`
 
+	// Enum to indicate the mode of resource creation.
+	CreateMode *CreateMode `json:"createMode,omitempty"`
+
+	// Parameters to indicate the information about the restore
+	RestoreParameters *ResourceRestoreParameters `json:"restoreParameters,omitempty"`
+
 	// A system generated property that specifies the addressable path of the users resource.
 	Users *string `json:"_users,omitempty"`
 
@@ -5371,6 +5459,12 @@ type SQLDatabaseListResult struct {
 type SQLDatabaseResource struct {
 	// REQUIRED; Name of the Cosmos DB SQL database
 	ID *string `json:"id,omitempty"`
+
+	// Enum to indicate the mode of resource creation.
+	CreateMode *CreateMode `json:"createMode,omitempty"`
+
+	// Parameters to indicate the information about the restore
+	RestoreParameters *ResourceRestoreParameters `json:"restoreParameters,omitempty"`
 }
 
 // SQLDedicatedGatewayRegionalServiceResource - Resource for a regional service location.
@@ -5590,6 +5684,20 @@ type SQLResourcesClientBeginSQLContainerRedistributeThroughputOptions struct {
 // SQLResourcesClientBeginSQLContainerRetrieveThroughputDistributionOptions contains the optional parameters for the SQLResourcesClient.BeginSQLContainerRetrieveThroughputDistribution
 // method.
 type SQLResourcesClientBeginSQLContainerRetrieveThroughputDistributionOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
+// SQLResourcesClientBeginSQLDatabaseRedistributeThroughputOptions contains the optional parameters for the SQLResourcesClient.BeginSQLDatabaseRedistributeThroughput
+// method.
+type SQLResourcesClientBeginSQLDatabaseRedistributeThroughputOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
+// SQLResourcesClientBeginSQLDatabaseRetrieveThroughputDistributionOptions contains the optional parameters for the SQLResourcesClient.BeginSQLDatabaseRetrieveThroughputDistribution
+// method.
+type SQLResourcesClientBeginSQLDatabaseRetrieveThroughputDistributionOptions struct {
 	// Resumes the LRO from the provided token.
 	ResumeToken string
 }